extracted NestedRVItemDecoration from nested rv example
